requirements.txt
requests
websocket-client
flask
flask-socketio
pytest
